Enable private plan switch to public plan
